+++ src/pkg_bulid/constructor/__init__.py
@@ -1,14 +1,14 @@
 """Constructor module for building and compressing packages."""
 
-from .build import Bulider, check_if_all_config_files_exists
+from .build import Builder, check_if_all_config_files_exist
 from .compresser import compress, parse_gitignore
 from .errors import ConfigFileNotFound, CompressionError, BuildError
 
 __all__ = [
-    "Bulider",
+    "Builder",
     "compress",
     "parse_gitignore",
-    "check_if_all_config_files_exists",
+    "check_if_all_config_files_exist",
     "ConfigFileNotFound",
     "CompressionError",
     "BuildError",