+++ src/pkg_server/database/config.py
@@ -2,10 +2,10 @@ from flask_sqlalchemy import SQLAlchemy
 from sqlalchemy.orm import DeclarativeBase
 
 
-db = SQLAlchemy()
+class Base(DeclarativeBase): ...
+
+db = SQLAlchemy(model_class=Base)
 
 class Config(object):
-    SQLALCHEMY_DATABASE_URI = "sqlite:///my_database.db"
+    SQLALCHEMY_DATABASE_URI = "sqlite:///pkg_database.db"
     SQLALCHEMY_TRACK_MODIFICATIONS = False
-
-class Base(DeclarativeBase): ...